samples/reasoning_engine/requirements.txt
google-cloud-aiplatform[reasoningengine,langchain]==1.57.0
langchain-google-alloydb-pg==0.4.1
langchain-google-vertexai==1.0.4